% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_report.R
\name{html_report}
\alias{html_report}
\title{Custom RMarkdown Theme}
\usage{
html_report(toc = TRUE, theme = "flatly", ...)
}
\arguments{
\item{...}{Arguments passed to the output format function.}
}
\value{
A modified HTML or PDF output format with the custom theme.
}
\description{
Custom RMarkdown Theme
}
